Problem: Client tests were not fully offline
Solution: Mock responses; refactor to reduce duplication
