+++ tests/unit/conftest.py
@@ -2,6 +2,7 @@ import json
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 from typing import Any, Callable, Dict, List
+from unittest.mock import AsyncMock, MagicMock
 
 import pytest as pytest
 from aleph_message.models import AggregateMessage, AlephMessage, PostMessage
@@ -10,7 +11,9 @@ import aleph.sdk.chains.ethereum as ethereum
 import aleph.sdk.chains.sol as solana
 import aleph.sdk.chains.substrate as substrate
 import aleph.sdk.chains.tezos as tezos
+from aleph.sdk import AlephHttpClient, AuthenticatedAlephHttpClient
 from aleph.sdk.chains.common import get_fallback_private_key
+from aleph.sdk.types import Account
 
 
 @pytest.fixture
@@ -112,6 +115,12 @@ def aleph_messages() -> List[AlephMessage]:
 
 
 @pytest.fixture
+def json_post() -> dict:
+    with open(Path(__file__).parent / "post.json", "r") as f:
+        return json.load(f)
+
+
+@pytest.fixture
 def raw_messages_response(aleph_messages) -> Callable[[int], Dict[str, Any]]:
     return lambda page: {
         "messages": [message.dict() for message in aleph_messages]
@@ -122,3 +131,85 @@ def raw_messages_response(aleph_messages) -> Callable[[int], Dict[str, Any]]:
         "pagination_per_page": max(len(aleph_messages), 20),
         "pagination_total": len(aleph_messages) if page == 1 else 0,
     }
+
+
+@pytest.fixture
+def raw_posts_response(json_post) -> Callable[[int], Dict[str, Any]]:
+    return lambda page: {
+        "posts": [json_post] if int(page) == 1 else [],
+        "pagination_item": "posts",
+        "pagination_page": int(page),
+        "pagination_per_page": 1,
+        "pagination_total": 1 if page == 1 else 0,
+    }
+
+
+class MockResponse:
+    def __init__(self, sync: bool):
+        self.sync = sync
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
+        ...
+
+    async def raise_for_status(self):
+        ...
+
+    @property
+    def status(self):
+        return 200 if self.sync else 202
+
+    async def json(self):
+        message_status = "processed" if self.sync else "pending"
+        return {
+            "message_status": message_status,
+            "publication_status": {"status": "success", "failed": []},
+        }
+
+    async def text(self):
+        return json.dumps(await self.json())
+
+
+@pytest.fixture
+def mock_session_with_post_success(
+    ethereum_account: Account,
+) -> AuthenticatedAlephHttpClient:
+    http_session = AsyncMock()
+    http_session.post = MagicMock()
+    http_session.post.side_effect = lambda *args, **kwargs: MockResponse(
+        sync=kwargs.get("sync", False)
+    )
+
+    client = AuthenticatedAlephHttpClient(
+        account=ethereum_account, api_server="http://localhost"
+    )
+    client.http_session = http_session
+
+    return client
+
+
+def make_custom_mock_response(resp_json, status=200) -> MockResponse:
+    class CustomMockResponse(MockResponse):
+        async def json(self):
+            return resp_json
+
+        @property
+        def status(self):
+            return status
+
+    return CustomMockResponse(sync=True)
+
+
+def make_mock_get_session(get_return_value: Dict[str, Any]) -> AlephHttpClient:
+    class MockHttpSession(AsyncMock):
+        def get(self, *_args, **_kwargs):
+            return make_custom_mock_response(get_return_value)
+
+    http_session = MockHttpSession()
+
+    client = AlephHttpClient(api_server="http://localhost")
+    client.http_session = http_session
+
+    return client
